--- docs/source/concepts/articleretrieval.rst
@@ -1,4 +1,0 @@
-Article Retrieval
-=================
-
-Article Retrieval operations tutorial
