+++ docs/source/factiva.analytics/article_fetcher.rst
@@ -1,18 +1,18 @@
-Article Retrieval Service
-#########################
+Article API Service
+###################
 When enabled along with all other Factiva Analytics APIs, this service
 allows to retrieve content for display purposes when end users need to
 read the underlying content from a calculated score or derived datapoint
 after processing Snapshots or Streams news articles.
 
-ArticleRetrieval
+ArticleFetcher
 ----------------
 
-.. autoclass:: factiva.analytics.article_retrieval.article_retrieval.ArticleRetrieval
+.. autoclass:: factiva.analytics.article_fetcher.article_fetcher.ArticleFetcher
    :members:
 
 UIArticle
 ---------
 
-.. autoclass:: factiva.analytics.article_retrieval.article_retrieval.UIArticle
+.. autoclass:: factiva.analytics.article_fetcher.article_fetcher.UIArticle
    :members:
