+++ setup.py
@@ -23,7 +23,7 @@ setup(
 
     # Warning: the folder 'factiva' should NOT have an __init__.py file to avoid conflicts with the same namespace across other packages
     package_dir={'': 'src'},
-    packages=['factiva.analytics', 'factiva.analytics.article_retrieval', 'factiva.analytics.auth',
+    packages=['factiva.analytics', 'factiva.analytics.article_fetcher', 'factiva.analytics.auth',
               'factiva.analytics.common', 'factiva.analytics.taxonomy', 'factiva.analytics.snapshots',
               'factiva.analytics.streams', 'factiva.analytics.integration'
             #   , 'factiva.analytics.lists',
