+++ src/factiva/analytics/article_fetcher/__init__.py
@@ -1,0 +1,7 @@
+"""
+    Define methods and properties for the Article API Service
+"""
+
+__all__ = ['ArticleFetcher', 'UIArticle']
+
+from .article_fetcher import ArticleFetcher, UIArticle
