--- src/factiva/analytics/article_retrieval/__init__.py
@@ -1,7 +1,0 @@
-"""
-    Define methods and properties for the Article Retrieval Service
-"""
-
-__all__ = ['ArticleRetrieval', 'UIArticle']
-
-from .article_retrieval import ArticleRetrieval, UIArticle
