+++ src/factiva/analytics/common/const.py
@@ -76,8 +76,8 @@ DOC_COUNT_EXCEEDED = "DOC_COUNT_EXCEEDED"
 CHECK_EXCEEDED_WAIT_SPACING = 300
 PUBSUB_MESSAGES_WAIT_SPACING = 10
 
-# Article Retrieval
-API_RETRIEVAL_ENDPOINT_BASEURL = '/content/refs'
+# Article API (Article Fetch)
+API_ARTICLE_ENDPOINT_BASEURL = '/content/refs'
 
 # API STATES
 API_JOB_CREATED_STATE = 'JOB_CREATED'
