+++ test/article_fetcher/test_articlefetcher.py
@@ -1,9 +1,9 @@
 """
-    Tests for the ArticleRetrieval module
+    Tests for the ArticleFetcher module
 """
 import pytest
 import time
-from factiva.analytics import OAuthUser, ArticleRetrieval, UIArticle
+from factiva.analytics import OAuthUser, ArticleFetcher, UIArticle
 from factiva.analytics.common import config, const
 
 GITHUB_CI = config.load_environment_value('CI', False)
@@ -28,17 +28,17 @@ def _assert_uiarticle_values(uiarticle: UIArticle):
     assert uiarticle.source_name == 'The Wall Street Journal Online'
 
 
-def test_article_retrieval_env_user():
+def test_article_fetcher_env_user():
     """"
     Creates the object using the ENV variable and request the article content to the API service
     """
     time.sleep(const.TEST_REQUEST_SPACING_SECONDS)
-    ar = ArticleRetrieval()
-    article = ar.retrieve_single_article(ARTICLE_ID)
+    ar = ArticleFetcher()
+    article = ar.fetch_single_article(ARTICLE_ID)
     _assert_uiarticle_values(article)
 
 
-def test_article_retrieval_params_user():
+def test_article_fetcher_params_user():
     """
     Creates the object using the passed params and request the article content to the API service
     """
@@ -46,7 +46,7 @@ def test_article_retrieval_params_user():
     o = OAuthUser(client_id=FACTIVA_CLIENTID,
                   username=FACTIVA_USERNAME,
                   password=FACTIVA_PASSWORD)
-    ar = ArticleRetrieval(oauth_user=o)
-    article = ar.retrieve_single_article(ARTICLE_ID)
+    ar = ArticleFetcher(oauth_user=o)
+    article = ar.fetch_single_article(ARTICLE_ID)
     _assert_uiarticle_values(article)
